requirements.txt
python-dotenv
livekit[rtc,agents]
requests
livekit-plugins[cartesia,deepgram,silero,groq,noise_cancellation]
livekit-turn-detector[multilingual]
